Refactor startup code so majority of charm init happens in one place
This means that all chares are created at one time, and all of the
globals have been set at this point. More cleanup is needed still, and
most of this can also probably be merged into tw_run in some way so that
there is only one place where we have to start the charm scheduler
(other than the very beginning for main chare).

Change-Id: I04644214422126c4eb7156b5f6f5ba26237b4701