Fix leash triggers for asynchronous reductions
In cases where the synchronous GVT is used but with async reductions,
the leash trigger was still based off the previous GVT. This means the
leash trigger will almost always immediately trigger again. However this
trigger was ignored because a GVT computation was already in progress,
and therefore the simulation could get arbitrarily far ahead.